app/requirements.txt
scikit-learn==0.23.2
nltk==3.5
spacy==2.3.2
textacy==0.10.0
pytextrank==2.0.2
pandas==1.1.0
joblib==0.16.0
wasabi==0.7.1
elasticsearch>=7.0.0,<8.0.0
elasticsearch-dsl>=7.0.0,<8.0.0
fastapi==0.61.0
uvicorn==0.11.8
gunicorn==20.0.4
apache-libcloud==3.2.0
boto3==1.15.10
python-dotenv==0.14.0
psycopg2==2.8.6
sqlalchemy==1.3.20
python-multipart==0.0.5
python-box==5.2.0
aiofiles==0.6.0
